Move NewTwitchApi into its own top-level namespace.
