+++ bin/new-api-cli-test-client.php
@@ -1,8 +1,9 @@
+#!/usr/bin/env php
 <?php
 
 require __DIR__ . '/../vendor/autoload.php';
 
-use TwitchApi\NewApi\CLI\CLIClient;
+use NewTwitchApi\CLI\CLIClient;
 
 try {
     (new CLIClient($argv))->run();
