+++ src/NewTwitchApi/Auth/AuthGuzzleClient.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace TwitchApi\NewApi\Auth;
+namespace NewTwitchApi\Auth;
 
 use GuzzleHttp\Client;
