+++ src/NewTwitchApi/Auth/Oauth.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace TwitchApi\NewApi\Auth;
+namespace NewTwitchApi\Auth;
 
 use GuzzleHttp\Client;
 use Psr\Http\Message\ResponseInterface;
