+++ src/NewTwitchApi/HelixGuzzleClient.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace TwitchApi\NewApi;
+namespace NewTwitchApi;
 
 use GuzzleHttp\Client;
